chore: move medimage module
