+++ src/imgtools/coretypes/__init__.py
@@ -1,4 +1,5 @@
 from .box import BoxPadMethod, RegionBox
+from .base_medimage import MedImage
 from .spatial_types.coord_types import Coordinate3D, Size3D, Spacing3D
 
-__all__ = ["Coordinate3D", "Size3D", "Spacing3D", "RegionBox", "BoxPadMethod"]
+__all__ = ["MedImage", "Coordinate3D", "Size3D", "Spacing3D", "RegionBox", "BoxPadMethod"]
