+++ src/imgtools/coretypes/base_medimage.py
