+++ src/imgtools/coretypes/imagetypes/__init__.py
@@ -1,6 +1,5 @@
-from .med_image import MedImage  # noqa
 from .dose import Dose
 from .pet import PET
 from .scan import Scan
 
-__all__ = ["MedImage", "Dose", "PET", "Scan"]
+__all__ = ["Dose", "PET", "Scan"]
