+++ src/imgtools/coretypes/imagetypes/scan.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Dict
 
-from imgtools.coretypes.imagetypes import MedImage
+from imgtools.coretypes import MedImage
 from imgtools.io.readers import read_dicom_series
 from imgtools.loggers import logger
